refactor mow lookup
