+++ src/fsd/1-pages/learn-mow/index.ts
@@ -1,0 +1,2 @@
+export { mowLookupDesktopLazyRoute } from './mow-lookup.route';
+export { mowLookupMenuItem } from './mow-lookup.menu-item';
