+++ src/fsd/1-pages/learn-mow/mow-lookup.menu-item.tsx
@@ -1,0 +1,6 @@
+﻿import MilitaryTechIcon from '@mui/icons-material/MilitaryTech';
+import React from 'react';
+
+import { MenuItemTP } from '@/fsd/5-shared/ui';
+
+export const mowLookupMenuItem = new MenuItemTP('MoW Lookup', <MilitaryTechIcon />, '/learn/mowLookup');
