+++ src/fsd/1-pages/learn-mow/mow-materials-table.tsx
@@ -2,15 +2,11 @@
 import { AgGridReact } from 'ag-grid-react';
 import React, { useMemo } from 'react';
 
-import { BadgeImage } from 'src/v2/components/images/badge-image';
-import { ComponentImage } from 'src/v2/components/images/component-image';
-import { ForgeBadgeImage } from 'src/v2/components/images/forge-badge-image';
+import { numberToThousandsString } from '@/fsd/5-shared/lib';
+import { BadgeImage, ComponentImage, ForgeBadgeImage } from '@/fsd/5-shared/ui/icons';
 
-import { numberToThousandsString } from '@/fsd/5-shared/lib/number-to-thousands-string';
-
-import { UpgradeImage } from '@/fsd/4-entities/upgrade/upgrade-image';
-
-import { IMowLevelMaterials } from 'src/v2/features/lookup/lookup.models';
+import { IMowLevelMaterials } from '@/fsd/4-entities/mow';
+import { UpgradeImage } from '@/fsd/4-entities/upgrade';
 
 interface Props {
     rows: IMowLevelMaterials[];
