+++ src/fsd/4-entities/mow/data/index.ts
@@ -1,0 +1,9 @@
+import { IMowLevelUpgrade, IMowLevelUpgradesDictionary, IMowStatic } from '../model';
+
+import mowLvlUpCommonJson from './mow-lvl-up-common.json';
+import mowUpgradesJson from './mows-upgrades.json';
+import mowsJson from './mows.json';
+
+export const mowsData = mowsJson as IMowStatic[];
+export const mowLevelUpCommonData: IMowLevelUpgrade[] = mowLvlUpCommonJson;
+export const mowUpgradesData: IMowLevelUpgradesDictionary = mowUpgradesJson;
