+++ src/fsd/4-entities/mow/index.ts
@@ -1,2 +1,6 @@
-export type { IMow, IMowDb, IMowStatic } from './model';
+export type { IMow, IMowDb, IMowStatic, IMowLevelMaterials, IMowLevelUpgrade } from './model';
 export { MowTitle } from './mow-title';
+export { MowUpgrades } from './mow-upgrades';
+export { MowUpgradesUpdate } from './mow-upgrades-update';
+export { MowsService } from './mows.service';
+export { mowsData } from './data';
