+++ src/fsd/4-entities/mow/model.ts
@@ -1,5 +1,7 @@
 import { Alliance, DynamicProps, Faction, Rarity, RarityStars, RarityString, UnitType } from '@/fsd/5-shared/model';
 
+import { IBaseUpgrade, ICraftedUpgrade } from '@/fsd/4-entities/upgrade/@x/mow';
+
 export interface IMowStatic {
     id: string;
     tacticusId: string;
@@ -9,6 +11,7 @@ export interface IMowStatic {
     fullName: string;
     releaseDate?: string;
     alliance: Alliance;
+    deployableAlliance: Alliance;
     faction: Faction;
     initialRarity: RarityString;
 }
@@ -27,4 +30,36 @@ export interface IMow extends IMowStatic, IMowDb, DynamicProps {
     portraitIcon: string;
     badgeIcon: string;
     unitType: UnitType.mow;
+}
+
+export interface IMowLevelUpgrade {
+    lvl: number;
+    components: number;
+    gold: number;
+    badges: number;
+    salvage?: number;
+    forgeBadges?: number;
+}
+
+export interface IMowLevelUpgrades {
+    lvl: number;
+    primary: string[];
+    secondary?: string[];
+}
+
+export type IMowLevelUpgradesDictionary = Record<string, Array<IMowLevelUpgrades>>;
+
+export interface IMowLevelMaterials {
+    mowId: string;
+    mowLabel: string;
+    mowAlliance: Alliance;
+    level: number;
+    components: number;
+    gold: number;
+    badges: number;
+    salvage: number;
+    forgeBadges: number;
+    rarity: Rarity;
+    primaryUpgrades: Array<IBaseUpgrade | ICraftedUpgrade>;
+    secondaryUpgrades: Array<IBaseUpgrade | ICraftedUpgrade>;
 }
