+++ src/fsd/4-entities/upgrade/@x/mow.ts
@@ -1,0 +1,3 @@
+export type { IBaseUpgrade, ICraftedUpgrade } from '../model';
+export { UpgradesService } from '../upgrades.service';
+export { UpgradeImage } from '../upgrade-image';
