+++ src/fsd/5-shared/ui/icons/forge-badge-image.tsx
@@ -1,8 +1,8 @@
 ﻿import React from 'react';
 
-import { getImageUrl } from 'src/shared-logic/functions';
-
 import { Rarity } from '@/fsd/5-shared/model';
+
+import { getImageUrl } from '../get-image-url';
 
 export const ForgeBadgeImage = ({ rarity, size = 'medium' }: { rarity: Rarity; size?: 'small' | 'medium' }) => {
     const sizePx = size === 'medium' ? 35 : 25;
