+++ src/fsd/5-shared/ui/icons/index.ts
@@ -5,4 +5,7 @@ export { RarityIcon } from './rarity.icon';
 export { StarsIcon } from './stars.icon';
 export { UnitShardIcon } from './unit-shard.icon';
 export { DamageIcon } from './damage-icon';
+export { BadgeImage } from './badge-image';
+export { ComponentImage } from './component-image';
+export { ForgeBadgeImage } from './forge-badge-image';
 export { pooEmoji, starEmoji } from './emojies';
