+++ src/fsd/5-shared/ui/input/index.ts
@@ -1,2 +1,3 @@
 export { Checkbox, type CheckboxProps, CheckboxGroup, type CheckboxGroupProps } from './checkbox';
 export { TextField, type TextFieldProps } from './text-field';
+export { NumberInput } from './number-input';
