+++ src/fsd/5-shared/ui/input/number-input.tsx
