+++ src/routes/goals/goals-table.tsx
@@ -31,7 +31,8 @@ import {
 } from 'src/v2/features/goals/goals.models';
 import { ShardsService } from 'src/v2/features/goals/shards.service';
 import { XpTotal } from 'src/v2/features/goals/xp-total';
-import { MowMaterialsTotal } from 'src/v2/features/lookup/mow-materials-total';
+
+import { MowMaterialsTotal } from '@/fsd/1-pages/learn-mow/mow-materials-total';
 
 interface Props {
     rows: CharacterRaidGoalSelect[];
