+++ src/shared-components/le-progress-overview-missions.tsx
@@ -1,7 +1,7 @@
 ﻿import { FormControlLabel, Switch } from '@mui/material';
 import React from 'react';
 
-import { NumberInput } from 'src/v2/components/inputs/number-input';
+import { NumberInput } from '@/fsd/5-shared/ui/input/number-input';
 
 import { ILreOccurrenceProgress } from 'src/v2/features/lre/lre.models';
