+++ src/v2/features/guides/components/incursion-modes-filter.tsx
@@ -1,8 +1,9 @@
 ﻿import React from 'react';
 
 import { MultipleSelect } from 'src/v2/components/inputs/multiple-select';
-import mowsData from 'src/v2/data/mows.json';
 import { IMenuOption } from 'src/v2/models/menu-option';
+
+import { mowsData } from '@/fsd/4-entities/mow';
 
 interface Props {
     selectedModes: string[];
