--- src/v2/pages/mow-lookup/mow-lookup.menu-item.tsx
@@ -1,6 +1,0 @@
-﻿import MilitaryTechIcon from '@mui/icons-material/MilitaryTech';
-import React from 'react';
-
-import { MenuItem } from 'src/v2/models/menu-item';
-
-export const mowLookupMenuItem = new MenuItem('MoW Lookup', <MilitaryTechIcon />, '/learn/mowLookup');
